requirements.txt
discord.py
PlexAPI
requests
aiohttp
python-dotenv
musicbrainzngs
schedule
